feat: move fetching of habits to a factory and add an addHabit method
This will make it easier to switch out our API to using Firestore or another system later on instead of JSON.
