+++ src/api/habits/adapters/index.ts
@@ -1,0 +1,3 @@
+import jsonHabitServiceFactory from './json';
+
+export default jsonHabitServiceFactory();
