+++ src/api/habits/index.ts
@@ -1,0 +1,3 @@
+import habitFactory from './adapters';
+
+export default habitFactory;
