+++ src/api/habits/types.ts
@@ -1,0 +1,10 @@
+import { GoalWithHabitHistory } from '../../types/habits';
+
+export type HabitService = {
+  save?: () => {};
+  retrieve?: () => {};
+  get?: () => {};
+  set?: () => {};
+  addHabit: () => void;
+  getHabits: () => Promise<GoalWithHabitHistory[]>;
+};
